Get a Node struct
