+++ rust/examples/01_the_switch_and_the_light/light.rs
@@ -37,8 +37,8 @@ pub fn main() {
     eprintln!("Connected to Arete control plane");
 
     // Register this node and its context with the control plane
-    let _system = client.system().unwrap();
-    client.add_node(NODE_ID, NODE_NAME, false, None).unwrap();
+    let system = client.system().unwrap();
+    let _node = system.node(NODE_ID, NODE_NAME, false, None).unwrap();
     eprintln!("Registered as node {NODE_ID}");
     client.add_context(NODE_ID, CONTEXT_ID, CONTEXT_NAME).unwrap();
     eprintln!("Registered context {CONTEXT_ID} for node {NODE_ID}");
